requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py310ha9d4c09_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py310h9ce1e76_2
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
focalloss=1.0.0=dev_0
fonttools=4.40.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py310h06a4308_0
imutils=0.5.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.14.0=pyh41d4057_0
ipywidgets=8.0.7=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=4.12.0=py310hff52083_0
jupyterlab-widgets=3.0.8=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.0.149=0
libcurand=10.3.3.53=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
numexpr=2.8.4=py310h85018f9_1
numpy=1.25.0=py310h5f9d8c6_0
numpy-base=1.25.0=py310hb5e798b_0
opencv-python=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.9=h7f8727e_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py310h1128e8f_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.1.2=py310h06a4308_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py310h06a4308_0
pyparsing=3.1.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.11=h955ad1f_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.13.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyzmq=25.1.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
requests=2.29.0=py310h06a4308_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.13.1=py310_cu117
torchvision=0.14.1=py310_cu117
tornado=6.1=py310h5764c6d_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.6.3=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py310h06a4308_0
widgetsnbextension=4.0.8=pypi_0
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
